refactor(cli): split tsz binary entrypoint into command and execution modules
`actual_main` previously interleaved argument validation, command dispatch,
and execution side effects in a single ~300-line path, coupling parsing to
execution and pushing `bin/tsz.rs` toward the 2000-line ceiling.

Introduce a clean parse -> execute boundary:

- Model the dispatch decision as a `Command` enum produced by
  `select_command(&mut CliArgs, &Path)`. It performs the same ordered
  validation/normalization tsc does (exiting in place on invalid argument
  combinations) and returns the action to run, so `actual_main` reads as a
  match-based orchestrator.
- Move the `--init` tsconfig-template machinery into `bin/tsz/init.rs`.
- Move the compile execution path (tracer setup, `driver::compile`, result
  list/diagnostic printing, exit-code computation) into `bin/tsz/run.rs`.
- Extract `validate_locale_or_exit` and a shared `print_no_input_help_and_exit`
  helper that de-duplicates three identical no-input help blocks.

Pure restructuring: process exit codes, stdout/stderr ordering, and all
diagnostic text are unchanged. `bin/tsz.rs` drops from 1778 to ~1010 lines.
